backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.1
pydantic[email]
airtable-python-wrapper==0.15.3
pyairtable
sendgrid==6.12.5
# PostgreSQL (invoices / billing)
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
# Auth: password hashing & JWT
passlib[argon2]>=1.7.4
pyjwt>=2.8.0
